+++ blog-backend/blog-data/src/main/java/com/donts/blog/entity/JobLog.java
@@ -22,12 +22,12 @@ public class JobLog implements Serializable {
      * 任务日志ID
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 任务ID
      */
-    private Integer jobId;
+    private Long jobId;
 
     /**
      * 任务名称
